+++ pyobjc-framework-Cocoa/PyObjCTest/test_nssplitviewitem.py
@@ -19,13 +19,6 @@ class TestNSSplitViewItem(TestCase):
         )
         self.assertEqual(AppKit.NSSplitViewItemCollapseBehaviorUseConstraints, 3)
 
-    @min_os_level("10.11")
-    def testConstants10_11(self):
-        self.assertIsInstance(AppKit.NSSplitViewItemUnspecifiedDimension, float)
-
-        self.assertResultIsBOOL(AppKit.NSSplitViewItem.isSpringLoaded)
-        self.assertArgIsBOOL(AppKit.NSSplitViewItem.setSpringLoaded_, 0)
-
     @min_os_level("10.10")
     def testMethods10_10(self):
         self.assertResultIsBOOL(AppKit.NSSplitViewItem.isCollapsed)
@@ -33,3 +26,15 @@ class TestNSSplitViewItem(TestCase):
 
         self.assertResultIsBOOL(AppKit.NSSplitViewItem.canCollapse)
         self.assertArgIsBOOL(AppKit.NSSplitViewItem.setCanCollapse_, 0)
+
+    @min_os_level("10.11")
+    def testMethods10_11(self):
+        self.assertIsInstance(AppKit.NSSplitViewItemUnspecifiedDimension, float)
+
+        self.assertResultIsBOOL(AppKit.NSSplitViewItem.isSpringLoaded)
+        self.assertArgIsBOOL(AppKit.NSSplitViewItem.setSpringLoaded_, 0)
+
+    @min_os_level("10.16")
+    def testMethods10_16(self):
+        self.assertResultIsBOOL(AppKit.NSSplitViewItem.allowsFullHeightLayout)
+        self.assertArgIsBOOL(AppKit.NSSplitViewItem.setAllowsFullHeightLayout_, 0)