+++ pyobjc-framework-MLCompute/PyObjCTest/test_mlcembeddingdescriptor.py
@@ -1,0 +1,15 @@
+from PyObjCTools.TestSupport import TestCase
+
+import MLCompute
+
+
+class TestMLCEmbeddingDescriptor(TestCase):
+    def test_methods(self):
+        self.assertResultIsBOOL(
+            MLCompute.MLCEmbeddingDescriptor.scalesGradientByFrequency
+        )
+
+        self.assertArgIsBOOL(
+            MLCompute.MLCEmbeddingDescriptor.descriptorWithEmbeddingCount_embeddingDimension_paddingIndex_maximumNorm_pNorm_scalesGradientByFrequency_,  # noqa: B950
+            5,
+        )