+++ pyobjc-framework-MLCompute/PyObjCTest/test_mlcoptimizer.py
@@ -1,0 +1,8 @@
+from PyObjCTools.TestSupport import TestCase
+
+import MLCompute
+
+
+class TestMLCOptimizer(TestCase):
+    def test_methods(self):
+        self.assertResultIsBOOL(MLCompute.MLCOptimizer.appliesGradientClipping)