+++ pyobjc-framework-MLCompute/PyObjCTest/test_mlctensordescriptor.py
@@ -1,0 +1,8 @@
+from PyObjCTools.TestSupport import TestCase
+
+import MLCompute
+
+
+class TestMLCTensorDescriptor(TestCase):
+    def test_methods(self):
+        self.assertResultIsBOOL(MLCompute.MLCTensorDescriptor.sortedSequences)